+++ addOns/client/src/main/java/org/zaproxy/addon/client/ClientIntegrationAPI.java
@@ -81,7 +81,9 @@ public class ClientIntegrationAPI extends ApiImplementor {
         return callbackUrl;
     }
 
-    private void handleReportObject(JSONObject json) {
+    private void handleReportObject(String jsonStr) {
+        LOGGER.debug("Got object: {}", jsonStr);
+        JSONObject json = JSONObject.fromObject(jsonStr);
         ReportedElement rnode = new ReportedElement(json);
         if (!"A".equals(rnode.getNodeName())) {
             // Dont add links - they flood the table
@@ -109,7 +111,9 @@ public class ClientIntegrationAPI extends ApiImplementor {
         }
     }
 
-    private void handleReportEvent(JSONObject json) {
+    private void handleReportEvent(String jsonStr) {
+        LOGGER.debug("Got event: {}", jsonStr);
+        JSONObject json = JSONObject.fromObject(jsonStr);
         ReportedEvent event = new ReportedEvent(json);
         if (event.getUrl() == null || !ExtensionClientIntegration.isApiUrl(event.getUrl())) {
             this.extension.addReportedObject(event);
@@ -121,43 +125,25 @@ public class ClientIntegrationAPI extends ApiImplementor {
 
     @Override
     public ApiResponse handleApiAction(String name, JSONObject params) throws ApiException {
-        JSONObject json;
-        switch (name) {
-            case ACTION_REPORT_OBJECT:
-                String objJson = this.getParam(params, PARAM_OBJECT_JSON, "");
-                LOGGER.debug("Got object: {}", objJson);
-                json = JSONObject.fromObject(objJson);
-                handleReportObject(json);
-                break;
+        try {
+            switch (name) {
+                case ACTION_REPORT_OBJECT -> handleReportObject(
+                        this.getParam(params, PARAM_OBJECT_JSON, ""));
 
-            case ACTION_REPORT_EVENT:
-                String eventJson = this.getParam(params, PARAM_EVENT_JSON, "");
-                LOGGER.debug("Got event: {}", eventJson);
-                json = JSONObject.fromObject(eventJson);
-                handleReportEvent(json);
-                break;
+                case ACTION_REPORT_EVENT -> handleReportEvent(
+                        this.getParam(params, PARAM_EVENT_JSON, ""));
 
-            case ACTION_REPORT_ZEST_STATEMENT:
-                String statementJson = this.getParam(params, PARAM_STATEMENT_JSON, "");
-                LOGGER.debug("Got script: {}", statementJson);
-                try {
-                    this.extension.addZestStatement(statementJson);
-                } catch (Exception e) {
-                    LOGGER.debug(e);
-                }
-                break;
-            case ACTION_REPORT_ZEST_SCRIPT:
-                String scriptJson = this.getParam(params, PARAM_SCRIPT_JSON, "");
-                LOGGER.debug("Got script: {}", scriptJson);
-                try {
-                    this.extension.addZestStatement(scriptJson);
-                } catch (Exception e) {
-                    LOGGER.debug(e);
-                }
-                break;
-
-            default:
-                throw new ApiException(ApiException.Type.BAD_ACTION);
+                case ACTION_REPORT_ZEST_STATEMENT -> this.extension.addZestStatement(
+                        this.getParam(params, PARAM_STATEMENT_JSON, ""));
+                case ACTION_REPORT_ZEST_SCRIPT -> this.extension.addZestStatement(
+                        this.getParam(params, PARAM_SCRIPT_JSON, ""));
+                default -> throw new ApiException(ApiException.Type.BAD_ACTION);
+            }
+        } catch (ApiException e) {
+            throw e;
+        } catch (Exception e) {
+            LOGGER.error(e.getMessage(), e);
+            throw new ApiException(ApiException.Type.INTERNAL_ERROR);
         }
 
         return ApiResponseElement.OK;
@@ -174,43 +160,30 @@ public class ClientIntegrationAPI extends ApiImplementor {
         return str;
     }
 
-    static JSONObject decodeParam(String body, String param) {
-        String str = decodeParamString(body, param);
-        return JSONObject.fromObject(str);
-    }
-
     @Override
     public String handleCallBack(HttpMessage msg) throws ApiException {
         if (HttpRequestHeader.POST.equals(msg.getRequestHeader().getMethod())) {
             String body = msg.getRequestBody().toString();
 
             if (body.startsWith(PARAM_OBJECT_JSON + "=")) {
-                JSONObject json = decodeParam(body, PARAM_OBJECT_JSON);
-                LOGGER.debug("Got object: {}", json);
-                handleReportObject(json);
-
+                handleReportObject(decodeParamString(body, PARAM_OBJECT_JSON));
             } else if (body.startsWith(PARAM_EVENT_JSON)) {
-                JSONObject json = decodeParam(body, PARAM_EVENT_JSON);
-                LOGGER.debug("Got event: {}", json);
-                handleReportEvent(json);
+                handleReportEvent(decodeParamString(body, PARAM_EVENT_JSON));
             } else if (body.startsWith(PARAM_STATEMENT_JSON)) {
                 try {
                     this.extension.addZestStatement(decodeParamString(body, PARAM_STATEMENT_JSON));
                 } catch (Exception e) {
-                    LOGGER.debug(e);
+                    LOGGER.error(e.getMessage(), e);
                 }
             } else if (body.startsWith(PARAM_SCRIPT_JSON)) {
                 try {
                     this.extension.addZestStatement(decodeParamString(body, PARAM_SCRIPT_JSON));
                 } catch (Exception e) {
-                    LOGGER.debug(e);
+                    LOGGER.error(e.getMessage(), e);
                 }
             }
-
-        } else {
-            // Will be accessed via a GET as part of the browser ext initiation
-            msg.setResponseBody(ApiResponseElement.OK.toJSON().toString());
         }
+        // Will be accessed via a GET as part of the browser ext initiation
         return "";
     }
 }