+++ addOns/client/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Extension not enabled when launched from ZAP.
+- Browser recording not enabled when launched from ZAP recorder.
 
 ## [0.11.0] - 2025-01-17
 ### Fixed